fix(bug): fix linked list collision implementation
The initial approach overwrites the initial value at the conflicting
address. It created a new linkedlist with the new value as the head. I
stored the current value in an intermediate variable and created a
linked list with the head node created from the initial value at the
address. Then I appended the new value to the linked list.